feat: show genres below each card of trending films
